Splash screen
The splash screen displays "Loading" until the application has loaded or
if the browser is determined to be unsupported. If the browser is
unsupported, an info screen is displayed. The browser support checking
is very rudimentary and takes into account a differing behavior of
dispatchEvent for IE compared to Chrome and Firefox.